refactored BoltzmannSelection Tests
